portfolio website fixes and structure changes
