+++ .env
@@ -1,0 +1,3 @@
+VITE_EMAILJS_SERVICE_ID='service_r4dxup6'
+VITE_EMAILJS_TEMPLATE_ID='template_lpo1zck'
+VITE_EMAILJS_PUBLIC_KEY='bgb4plZbiyJcv77go'
