+++ package.json
@@ -13,11 +13,14 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@emailjs/browser": "^4.4.1",
     "@radix-ui/react-label": "^2.1.1",
     "@radix-ui/react-slot": "^1.1.1",
+    "@sendgrid/mail": "^8.1.5",
     "animejs": "^4.0.1",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
+    "dotenv": "^17.2.2",
     "framer-motion": "^12.0.6",
     "lucide-react": "^0.474.0",
     "motion": "^12.2.0",
