+++ src/components/widgets/BackgroundMusic.tsx
