+++ src/components/widgets/PixelLoader.tsx
